// src/services/Question/action-types.ts
export enum questionsActionTypes {
  LOADING = 'LOADING_QUESTIONS',
  LOADED = 'LOADED_QUESTIONS',
  SET_LIST = 'SET_QUESTIONS_LIST',
  ADD = 'ADD',
  ADD_UNSAVED = 'ADD_UNSAVED_QUESTION',
  DELETE = 'DELETE_QUESTION',
  DELETE_MULTIPLE = 'DELETE_MULTIPLE_QUESTIONS',
  UPDATE_QUESTION = 'UPDATE_QUESTION',
}